% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e11.11.F}
\alias{e11.11.F}
\title{E11 frequency matrix}
\format{A matrix with 170288 rows and 11 columns.}
\usage{
e11.11.F
}
\description{
e11.11.F is a frequency matrix used in E11.
}
\keyword{datasets}
