% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e11.alleles}
\alias{e11.alleles}
\title{SNPs used in E11}
\format{A data frame with 170288 rows and 5 columns.}
\usage{
e11.alleles
}
\description{
e11.alleles is a description of all SNPs used in E11.
}
\keyword{datasets}
