% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br.R
\name{fFixBr}
\alias{fFixBr}
\title{block relaxation with fixed F}
\usage{
fFixBr(gnew, qnew, f, acc, max.iter = 3, tol = 1e-04, pubdata = NULL)
}
\arguments{
\item{gnew}{Integer which length is the number of SNPs used in calculation.}

\item{qnew}{Initial q used in calculation. Numeric. Sum(q) must be 1.}

\item{f}{Allele frequencies learned from the reference panels.}

\item{acc}{a logical value indicating whether use quasi-Newton accelerated BR or not.}

\item{max.iter}{If acc = T, max.iter must be set, the default is 3.}

\item{tol}{If acc = F, tolerance must be set, the default is 1e-4.}

\item{pubdata}{You can choose a public dataset here, E11 or K13. You also can use other public
dataset which is not in this package.}
}
\description{
block relaxation with fixed F
}
