% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{fFixEm}
\alias{fFixEm}
\title{EM for fixed F}
\usage{
fFixEm(gnew, qnew, f, acc, max.iter = 3, tol = 1e-04, pubdata = NULL)
}
\arguments{
\item{gnew}{Integer which length is the number of SNPs used in calculation.}

\item{qnew}{Initial q used in calculation. Numeric. Sum(q) must be 1.}

\item{f}{Allele frequencies which was learned from the reference panels.}

\item{acc}{a logical value indicating whether use quasi-Newton accelerated EM or not.}

\item{max.iter}{an integer. If acc is TRUE, the number of iterations must be set.}

\item{tol}{Tolerance. If acc is FALSE, tol must be set. The default is 1e-4.}

\item{pubdata}{You can choose a public dataset here, E11 or K13. You also can use other public
dataset which is not in this package.}
}
\value{
Estimation results of q and the loglikelihood value of each iteration.
}
\description{
This function is for projection analysis which is an efficient way
of estimating individual ancestry  with allele frequencies learned from the reference panels.
}
