% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global13.13.F}
\alias{global13.13.F}
\title{K13 frequency matrix}
\format{A matrix with 166255 rows and 13 columns.}
\usage{
global13.13.F
}
\description{
This data set comes from Dodecad ancestry project.
global.13.13.F is a frequency matrix used in K4.
}
\keyword{datasets}
