% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global13.alleles}
\alias{global13.alleles}
\title{SNPs used in K13}
\format{A data frame with 166255 rows and 5 columns.}
\usage{
global13.alleles
}
\description{
This data set comes from Dodecad ancestry project.
global13.alleles is a description of all SNPs used in K13.
}
\keyword{datasets}

