% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globe4.4.F}
\alias{globe4.4.F}
\title{K4 frequency matrix}
\format{A matrix with 166255 rows and 4 columns.}
\usage{
globe4.4.F
}
\description{
globe4.4.F is a frequency matrix used in K4.
}
\keyword{datasets}
