% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globe4.alleles}
\alias{globe4.alleles}
\title{SNPs used in K4}
\format{A data frame with 166255 rows and 5 columns.}
\usage{
globe4.alleles
}
\description{
globe4.alleles is a description of all SNPs used in K4.
}
\keyword{datasets}
