% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K12b.12.F}
\alias{K12b.12.F}
\title{K12b frequency matrix}
\format{A matrix with 166770 rows and 12 columns.}
\usage{
K12b.12.F
}
\description{
K12b.12.F is a frequency matrix used in K12b.
}
\keyword{datasets}
