% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K12b.alleles}
\alias{K12b.alleles}
\title{SNPs used in K12b}
\format{A data frame with 166770 rows and 5 columns.}
\usage{
K12b.alleles
}
\description{
K12b.alleles is a description of all SNPs used in K12b.
}
\keyword{datasets}
