% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K7b.7.F}
\alias{K7b.7.F}
\title{K7b frequency matrix}
\format{A matrix with 166770 rows and 7 columns.}
\usage{
K7b.7.F
}
\description{
K7b.alleles is a frequency matrix used in K7b.
}
\keyword{datasets}
