% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K7b.alleles}
\alias{K7b.alleles}
\title{SNPs used in K7b}
\format{A data frame with 166770 rows and 5 columns.}
\usage{
K7b.alleles
}
\description{
K7b.alleles is a description of all SNPs used in K7.
}
\keyword{datasets}
