% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrd.R
\name{tfrdpub}
\alias{tfrdpub}
\title{transfer raw genotype file for f fixed admixture.}
\usage{
tfrdpub(genotype, K, map, f)
}
\arguments{
\item{genotype}{A data frame contains your genotype information.}

\item{K}{the number of populations}

\item{map}{Index file, it should contain rsid, choromosome position,
major allele and minor allele information for both plus and minus strands.}

\item{f}{Frequency matrix}
}
\value{
A list contains g, q, f which can be used for calculation.
}
\description{
transfer raw genotype file for f fixed admixture.
}

