% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world9.9.F}
\alias{world9.9.F}
\title{World9 frequency matrix}
\format{A matrix with 170822 rows and 9 columns.}
\usage{
world9.9.F
}
\description{
world9.9.F is a frequency matrix used in World9.
}
\keyword{datasets}
