% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world9.alleles}
\alias{world9.alleles}
\title{SNPs used in World9}
\format{A data frame with 170822 rows and 5 columns.}
\usage{
world9.alleles
}
\description{
world9.alleles is a description of all SNPs used in World9.
}
\keyword{datasets}
