% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrd.R
\name{generateG}
\alias{generateG}
\title{Transfer ped file to genotype matrix}
\usage{
generateG(referenceped)
}
\arguments{
\item{referenceped}{ped file for your reference panel}
}
\value{
genotype
}
\description{
Transfer ped file to genotype matrix
}
