% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_raw_data.R
\name{generateG}
\alias{generateG}
\title{Transfer ped file to genotype matrix}
\usage{
generateG(rawped)
}
\arguments{
\item{rawped}{A data.frame. Standard ped format. Genotype should be transferred to 1,2,3,4 from A,C,G,T. 0 represents missing.
'-','_','I','D' should be replaced by 0 by yourself.}
}
\value{
genotype matrix
}
\description{
This function can be used to transfer a ped file to g matrix
}
