% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_raw_data.R
\name{initQF}
\alias{initQF}
\title{Initialize Q and F}
\usage{
initQF(g, pop = NULL, alpha = NULL, K = NULL, model)
}
\arguments{
\item{g}{genotype matrix}

\item{pop}{A data.frame. If you intend to do supervised learning, 
you must specify the ancestries of the reference individuals.}

\item{alpha}{Parameter for dirichlet distribution.
Vector of shape parameters, or matrix of shape
parameters corresponding to the number of draw.}

\item{K}{If you intend to do unsupervised learning,
set the number of populations you will use.}

\item{model}{Choose supervised or unsupervised learning.}
}
\value{
A list contains q and f matrix.
}
\description{
This function could help you initialize Q and F matrix conveniently especially when 
you intend to use supervised learning.
}
