% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radmixture.R
\docType{package}
\name{radmixture}
\alias{radmixture}
\alias{radmixture-package}
\title{radmixture}
\description{
radmixture is an R package for ancestry calculation. It provides both supervised and
unsupervised learning with several algorithms for researchers and DNA customers.
see README on \href{https://github.com/wegene-llc/radmixture}{GitHub}
}
