% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_raw_data.R
\name{tfrdpub}
\alias{tfrdpub}
\title{Transfer personal genotype raw data according public dataset}
\usage{
tfrdpub(genotype, K, map, f)
}
\arguments{
\item{genotype}{A data.frame contains your genotype information.}

\item{K}{The number of populations}

\item{map}{A data.frame, it should contain rsid, major allele and minor allele 
information for both plus and minus strands.}

\item{f}{Frequency matrix learned from reference panel.}
}
\value{
A list contains g, q, f which can be used for calculation.
}
\description{
Transfer personal genotype raw data to g matrix which the number of row is 1 and
the number of column is the number of SNPs used here.
}
